# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Sistema de Panificación
 * OpenAPI spec version: 0.1.0
 */
export interface Proveedor {
  id: number;
  nombre: string;
  contacto?: string;
  telefono?: string;
  email?: string;
  direccion?: string;
  notas?: string;
  creadoEn: string;
  actualizadoEn: string;
}

export interface CrearProveedorBody {
  nombre: string;
  contacto?: string;
  telefono?: string;
  email?: string;
  direccion?: string;
  notas?: string;
}

export interface ResumenProveedor {
  proveedorId: number;
  proveedorNombre: string;
  totalCompras: number;
  totalInvertido: number;
  totalProductos: number;
  ultimaCompra?: string;
}

export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MensajeResponse {
  mensaje: string;
}

export interface Producto {
  id: number;
  codigo: string;
  nombre: string;
  descripcion: string;
  precioVenta: number;
  precioCosto: number;
  stock: number;
  stockMinimo: number;
  unidad: string;
  proveedorId?: number;
  proveedorNombre?: string;
  creadoEn: string;
  actualizadoEn: string;
}

export interface CrearProductoBody {
  codigo: string;
  nombre: string;
  descripcion: string;
  precioVenta: number;
  precioCosto: number;
  stock: number;
  stockMinimo: number;
  unidad: string;
  proveedorId?: number;
}

export interface ActualizarProductoBody {
  nombre?: string;
  descripcion?: string;
  precioVenta?: number;
  precioCosto?: number;
  stock?: number;
  stockMinimo?: number;
  unidad?: string;
  proveedorId?: number;
}

export interface ItemVenta {
  id: number;
  ventaId: number;
  productoCodigo: string;
  productoNombre: string;
  cantidad: number;
  precioUnitario: number;
  precioCosto: number;
  subtotal: number;
}

export interface Venta {
  id: number;
  vendedor: string;
  fecha: string;
  total: number;
  ganancia: number;
  items: ItemVenta[];
}

export interface ItemVentaInput {
  productoCodigo: string;
  cantidad: number;
}

export interface RegistrarVentaBody {
  vendedor: string;
  items: ItemVentaInput[];
}

export interface ItemCompra {
  id: number;
  compraId: number;
  productoCodigo: string;
  productoNombre: string;
  cantidad: number;
  precioCosto: number;
  subtotal: number;
}

export interface Compra {
  id: number;
  proveedor: string;
  fecha: string;
  totalInvertido: number;
  notas?: string;
  items: ItemCompra[];
}

export interface ItemCompraInput {
  productoCodigo: string;
  cantidad: number;
  precioCosto: number;
}

export interface RegistrarCompraBody {
  proveedor: string;
  notas?: string;
  items: ItemCompraInput[];
}

export interface ResumenFinanciero {
  totalVentas: number;
  totalInvertido: number;
  gananciaTotal: number;
  cantidadVentas: number;
  cantidadCompras: number;
  margenPromedio: number;
}

export interface TopProducto {
  codigo: string;
  nombre: string;
  cantidadVendida: number;
  ingresos: number;
  ganancia: number;
}

export interface VentaPorDia {
  fecha: string;
  totalVentas: number;
  cantidadVentas: number;
  ganancia: number;
}

export interface ReporteVendedor {
  vendedor: string;
  cantidadVentas: number;
  totalVentas: number;
  gananciaGenerada: number;
}

export type ListarVentasParams = {
  fechaDesde?: string;
  fechaHasta?: string;
  vendedor?: string;
};

export type ListarComprasParams = {
  fechaDesde?: string;
  fechaHasta?: string;
};

export type ObtenerResumenParams = {
  fechaDesde?: string;
  fechaHasta?: string;
};

export type ObtenerVentasPorDiaParams = {
  dias?: number;
};