# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Sistema de Panificación
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /proveedores:
    get:
      operationId: listarProveedores
      tags: [proveedores]
      summary: List all suppliers
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proveedor"
    post:
      operationId: crearProveedor
      tags: [proveedores]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrearProveedorBody"
      responses:
        "201":
          description: Created supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proveedor"
  /proveedores/{id}:
    get:
      operationId: obtenerProveedor
      tags: [proveedores]
      summary: Get supplier by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proveedor"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: actualizarProveedor
      tags: [proveedores]
      summary: Update a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrearProveedorBody"
      responses:
        "200":
          description: Updated supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proveedor"
    delete:
      operationId: eliminarProveedor
      tags: [proveedores]
      summary: Delete a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MensajeResponse"
  /proveedores/{id}/productos:
    get:
      operationId: obtenerProductosPorProveedor
      tags: [proveedores]
      summary: Get all products supplied by a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Products from supplier
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Producto"
  /proveedores/{id}/compras:
    get:
      operationId: obtenerComprasPorProveedor
      tags: [proveedores]
      summary: Get all purchases from a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchases from supplier
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Compra"
  /proveedores/{id}/resumen:
    get:
      operationId: obtenerResumenProveedor
      tags: [proveedores]
      summary: Get financial summary per supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumenProveedor"
  /productos:
    get:
      operationId: listarProductos
      tags: [productos]
      summary: List all products
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Producto"
    post:
      operationId: crearProducto
      tags: [productos]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrearProductoBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Producto"
  /productos/{codigo}:
    get:
      operationId: obtenerProducto
      tags: [productos]
      summary: Get product by code
      parameters:
        - name: codigo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Producto"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: actualizarProducto
      tags: [productos]
      summary: Update a product
      parameters:
        - name: codigo
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActualizarProductoBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Producto"
    delete:
      operationId: eliminarProducto
      tags: [productos]
      summary: Delete a product
      parameters:
        - name: codigo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MensajeResponse"
...
  /compras:
    get:
      operationId: listarCompras
      tags: [compras]
      summary: List all purchases
      parameters:
        - name: fechaDesde
          in: query
          required: false
          schema:
            type: string
        - name: fechaHasta
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of purchases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Compra"
    post:
      operationId: registrarCompra
      tags: [compras]
      summary: Register a purchase (stock replenishment)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrarCompraBody"
      responses:
        "201":
          description: Purchase registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Compra"
...
components:
  schemas:
    Proveedor:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        contacto:
          type: string
        telefono:
          type: string
        email:
          type: string
        direccion:
          type: string
        notas:
          type: string
        creadoEn:
          type: string
        actualizadoEn:
          type: string
      required: [id, nombre, creadoEn, actualizadoEn]
    CrearProveedorBody:
      type: object
      properties:
        nombre:
          type: string
        contacto:
          type: string
        telefono:
          type: string
        email:
          type: string
        direccion:
          type: string
        notas:
          type: string
      required: [nombre]
    ResumenProveedor:
      type: object
      properties:
        proveedorId:
          type: integer
        proveedorNombre:
          type: string
        totalCompras:
          type: integer
        totalInvertido:
          type: number
        totalProductos:
          type: integer
        ultimaCompra:
          type: string
      required: [proveedorId, proveedorNombre, totalCompras, totalInvertido, totalProductos]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MensajeResponse:
      type: object
      properties:
        mensaje:
          type: string
      required: [mensaje]
    Producto:
      type: object
      properties:
        id:
          type: integer
        codigo:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        precioVenta:
          type: number
        precioCosto:
          type: number
        stock:
          type: integer
        stockMinimo:
          type: integer
        unidad:
          type: string
        proveedorId:
          type: integer
        proveedorNombre:
          type: string
        creadoEn:
          type: string
        actualizadoEn:
          type: string
      required: [id, codigo, nombre, descripcion, precioVenta, precioCosto, stock, stockMinimo, unidad, creadoEn, actualizadoEn]
    CrearProductoBody:
      type: object
      properties:
        codigo:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        precioVenta:
          type: number
        precioCosto:
          type: number
        stock:
          type: integer
        stockMinimo:
          type: integer
        unidad:
          type: string
        proveedorId:
          type: integer
      required: [codigo, nombre, descripcion, precioVenta, precioCosto, stock, stockMinimo, unidad]
    ActualizarProductoBody:
      type: object
      properties:
        nombre:
          type: string
        descripcion:
          type: string
        precioVenta:
          type: number
        precioCosto:
          type: number
        stock:
          type: integer
        stockMinimo:
          type: integer
        unidad:
          type: string
        proveedorId:
          type: integer
...
    Compra:
      type: object
      properties:
        id:
          type: integer
        proveedor:
          type: string
        fecha:
          type: string
        totalInvertido:
          type: number
        notas:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ItemCompra"
      required: [id, proveedor, fecha, totalInvertido, items]
    ItemCompra:
      type: object
      properties:
        id:
          type: integer
        compraId:
          type: integer
        productoCodigo:
          type: string
        productoNombre:
          type: string
        cantidad:
          type: integer
        precioCosto:
          type: number
        subtotal:
          type: number
      required: [id, compraId, productoCodigo, productoNombre, cantidad, precioCosto, subtotal]
    RegistrarCompraBody:
      type: object
      properties:
        proveedor:
          type: string
        notas:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ItemCompraInput"
      required: [proveedor, items]
    ItemCompraInput:
      type: object
      properties:
        productoCodigo:
          type: string
        cantidad:
          type: integer
        precioCosto:
          type: number
      required: [productoCodigo, cantidad, precioCosto]
...